From: Thomas Dybdal Pedersen <thomasp85@gmail.com>

git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/FindMyFriends@114504 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/cdhit-common.cpp
@@ -2290,7 +2290,7 @@ size_t SequenceDB::MinimalMemory( int frag_no, int bsize, int T, const Options &
 
 	if(options.max_memory and options.max_memory < mem_need + 50*table ){
 		char msg[200];
-		sprintf( msg, "not enough memory, please set -M option greater than %zu\n", (50*table + mem_need/mega) );
+		sprintf( msg, "not enough memory, please set -M option greater than %lu\n", (unsigned long)(50*table + mem_need/mega) );
 		bomb_error(msg);
 	}
 	return mem_need;
